modules/result_processor/requirements.txt
fastapi
uvicorn
pandas
pymongo
python-dotenv
python-multipart
